+++ src/index_set.rs
@@ -1,4 +1,6 @@
+#![deny(clippy::undocumented_unsafe_blocks)]
 //! A fixed-capacity hash set where the iteration order is independent of the hash values.
+
 use core::{
     borrow::Borrow,
     fmt,
@@ -482,6 +484,28 @@ where
     /// The value may be any borrowed form of the set's value type, but `Hash` and `Eq` on the
     /// borrowed form must match those for the value type.
     ///
+    /// **NOTE**: This is equivalent to [`.swap_remove(key)`](IndexSet::swap_remove), replacing this
+    /// entry’s position with the last element, and it is deprecated in favor of calling that
+    /// explicitly.
+    #[deprecated(
+        note = "`remove` disrupts the set order -- use `swap_remove` for explicit behavior."
+    )]
+    pub fn remove<Q>(&mut self, value: &Q) -> bool
+    where
+        T: Borrow<Q>,
+        Q: ?Sized + Eq + Hash,
+    {
+        self.swap_remove(value)
+    }
+
+    /// Removes a value from the set. Returns `true` if the value was present in the set.
+    ///
+    /// The value may be any borrowed form of the set's value type, but `Hash` and `Eq` on the
+    /// borrowed form must match those for the value type.
+    ///
+    /// Like `Vec::swap_remove`, the value is removed by swapping it with the last element of the
+    /// map and popping it off. **This perturbs the position of what used to be the last element!**.
+    ///
     /// # Examples
     ///
     /// ```
@@ -490,15 +514,15 @@ where
     /// let mut set = FnvIndexSet::<_, 16>::new();
     ///
     /// set.insert(2).unwrap();
-    /// assert_eq!(set.remove(&2), true);
-    /// assert_eq!(set.remove(&2), false);
+    /// assert_eq!(set.swap_remove(&2), true);
+    /// assert_eq!(set.swap_remove(&2), false);
     /// ```
-    pub fn remove<Q>(&mut self, value: &Q) -> bool
+    pub fn swap_remove<Q>(&mut self, value: &Q) -> bool
     where
         T: Borrow<Q>,
         Q: ?Sized + Eq + Hash,
     {
-        self.map.remove(value).is_some()
+        self.map.swap_remove(value).is_some()
     }
 
     /// Retains only the elements specified by the predicate.